Learn a bit more about table views and models, some GUI updating logic
I finally started work on the GUI side with a QTableView. A example model
is added to MainWindow.py that generates a random number for a cell. This
allows for me to test that a table actually updates without using real data.

On the model side. The torrent can now accept a signal to emit every time
it succesfully updates it's data. This was tested using the above mentioned
method and some print statements.

There is still some more work to be done, but most of the ground work is layed
out.