Added more fields for display and fixed other minor things
Now we have a percentage of completion, the correct size of the downloaded bytes,
total upload and we can measure the rate of upload and download. These all work
as expected and was tested by moniotring a know working front end.

Now that Roles are understood the Model will only return the necessary values for the
display role.