Big changes, details bellow. Table ProgressBar Rendering and Torrent Monitoring
This should have been split but I got too exited with the work
so here it goes.

On the GUI Side:
The completion column is now turned into a progress bar. This entailed
the adding of a QStyledItemDelegate that custom draws the progress
bar item. It feeds of the data provided by the Table Model.

On the Comm Side:
The client can now handle torrent changes on the server side. This means
that removing a torrent no longer throws exceptions and adding torrents
works.

When torrents are added or removed the model now updates aswell.

Before dying each torrent send a print statement to keep track of
each torrent dying. For removed torrents I depend on the server
throwing an exception at me.

The TorrentManager now runs in it's own separate time thread at the
same refresh rate.